update to jest tests
